+++ pool-apps/pool/src/lib/channel_manager/mining_message_handler.rs
@@ -584,7 +584,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                                 channel_id,
                                 last_sequence_number: share_accounting.get_last_share_sequence_number(),
                                 new_submits_accepted_count: share_accounting.get_last_batch_accepted(),
-                                new_shares_sum: share_accounting.get_last_batch_work_sum() as u64,
+                                new_shares_sum: share_accounting.get_last_batch_work_sum(),
                             };
                             info!("SubmitSharesStandard: {} ✅", success);
                             messages.push((downstream_id, Mining::SubmitSharesSuccess(success)).into());
@@ -617,7 +617,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                             channel_id,
                             last_sequence_number: share_accounting.get_last_share_sequence_number(),
                             new_submits_accepted_count: share_accounting.get_last_batch_accepted(),
-                            new_shares_sum: share_accounting.get_last_batch_work_sum() as u64,
+                            new_shares_sum: share_accounting.get_last_batch_work_sum(),
                         };
                         messages.push((downstream_id, Mining::SubmitSharesSuccess(success)).into());
                     }
@@ -783,7 +783,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                                 channel_id,
                                 last_sequence_number: share_accounting.get_last_share_sequence_number(),
                                 new_submits_accepted_count: share_accounting.get_last_batch_accepted(),
-                                new_shares_sum: share_accounting.get_last_batch_work_sum() as u64,
+                                new_shares_sum: share_accounting.get_last_batch_work_sum(),
                             };
                             info!("SubmitSharesExtended: {} ✅", success);
                             messages.push((downstream_id, Mining::SubmitSharesSuccess(success)).into());
@@ -815,7 +815,7 @@ impl HandleMiningMessagesFromClientAsync for ChannelManager {
                             channel_id,
                             last_sequence_number: share_accounting.get_last_share_sequence_number(),
                             new_submits_accepted_count: share_accounting.get_last_batch_accepted(),
-                            new_shares_sum: share_accounting.get_last_batch_work_sum() as u64,
+                            new_shares_sum: share_accounting.get_last_batch_work_sum(),
                         };
                         messages.push((downstream_id, Mining::SubmitSharesSuccess(success)).into());
                     }
